[问题/需求描述]:适配新的传输协议 [解决方法]:适配新的传输协议 [影响范围]:无 [需求编号]:无

+++ domain/components/instrument_info.cc
@@ -27,7 +27,7 @@ std::string InstrumentInfo::GetExchange(const std::string &ins) {
   }
 }
 
-float InstrumentInfo::GetTickSize(const std::string &ins) {
+double InstrumentInfo::GetTickSize(const std::string &ins) {
   auto iter = info_map_.find(ins);
   if (iter != info_map_.end()) {
     return iter->second.ticksize;
